Bakend/requirements.txt
Django~=5.1.1
Django~=5.1.1
django-seed~=0.3.1
psycopg2-binary~=2.9.9
django-bootstrap5~=24.3
django-bootstrap-icons~=0.8.6
django-environ~=0.11.2
djangorestframework~=3.14.0
django-oauth-toolkit~=3.0.1
Pillow~=11.0.0
django-cors-headers